# specs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: 熊舍管家 MVP API
  version: 1.0.1
  summary: 熊舍管家 MVP 的 REST API 契约
  description: |
    面向 Flutter 客户端的模块化单体 REST API 草案。

    所有租户业务数据都由 Bearer 令牌中的认证上下文确定 owner_id；
    普通写请求体不接收 owner_id。状态迁移统一通过动作接口完成，
    客户端不得直接 PATCH state。

    除公开分享读取外，所有资源 ID 都先在认证 owner_id 范围内解析；
    不存在与跨 owner 资源统一返回 404，错误体不暴露其他账号的业务字段。

    写请求统一支持 Idempotency-Key。可并发编辑的资源通过 If-Match
    传入当前版本，响应同时返回 ETag 与资源 version。所有 date-time
    均以 UTC 传输，业务日期计算使用请求或当前熊舍的 IANA timezone。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.scolvpet.cn/v1
    description: 生产环境
  - url: https://api-staging.scolvpet.cn/v1
    description: 预发布环境
...
security:
  - bearerAuth: []
paths:
...
  /breeding-plans/{plan_id}/confirm-birth:
    parameters:
      - $ref: '#/components/parameters/PlanId'
    post:
      tags: [繁育, 窝次]
      operationId: confirmBirth
      summary: 确认产仔并建立窝次
      description: |
        仅允许 gestation 状态；同一计划只允许一个未撤销的生产事实。
        N>0 时幂等创建唯一有效 litter、初始数量流水及 N 条 pup_identity，并直接进入
        litter_nursing。N=0 时只记录无活仔生产结果并进入 no_litter_outcome，不创建
        litter、litter_count_event、pup_identity 或窝仔阶段任务。
        报喜卡或媒体任务失败不回滚产仔事实。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmBirthRequest'
            example:
              born_at: '2026-08-04T02:30:00Z'
              enclosure_id: 018f47a2-6450-7f54-8851-050a3d63846f
              initial_alive_count: 4
              initial_other_count: 1
              dam_condition:
                status: stable
                notes: 精神与进食正常
              temporary_code_prefix: L240804
              timezone: Asia/Shanghai
              outcome_reason: live_pups_observed
              notes: 发现 4 只活仔，另有 1 只死产
      responses:
        '201':
          description: 产仔事实、窝次和临时幼崽已创建
          headers:
            Location:
              $ref: '#/components/headers/Location'
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmBirthResponse'
              example:
                data:
                  result_type: live_litter
                  breeding_plan:
                    id: 018f47a2-4e63-7cc4-80b4-94a589917a36
                    owner_id: 018f47a2-31dc-7e1d-9f69-3b77e4c55d11
                    sire_id: 018f47a2-4fe7-7823-8e1b-672ee718b014
                    dam_id: 018f47a2-50bf-7489-ab99-8fcce45842a0
                    rule_version_id: 018f47a2-3e3b-7e40-9665-12cd57082cf0
                    state: litter_nursing
                    actual_birth_at: '2026-08-04T02:30:00Z'
                    version: 7
                    created_at: '2026-07-16T02:00:00Z'
                    updated_at: '2026-08-04T02:31:00Z'
                  litter:
                    id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                    owner_id: 018f47a2-31dc-7e1d-9f69-3b77e4c55d11
                    origin: breeding
                    code: SY-L-20260804-01
                    breeding_plan_id: 018f47a2-4e63-7cc4-80b4-94a589917a36
                    sire_id: 018f47a2-4fe7-7823-8e1b-672ee718b014
                    dam_id: 018f47a2-50bf-7489-ab99-8fcce45842a0
                    born_at: '2026-08-04T02:30:00Z'
                    initial_alive_count: 4
                    initial_other_count: 1
                    current_managed_count: 4
                    state: nursing
                    enclosure_id: 018f47a2-6450-7f54-8851-050a3d63846f
                    dam_condition:
                      status: stable
                      notes: 精神与进食正常
                    version: 1
                    created_at: '2026-08-04T02:31:00Z'
                    updated_at: '2026-08-04T02:31:00Z'
                  pup_identity_count: 4
                  pup_identities:
                    - id: 018f47a2-73b3-762e-8498-07b13dc3b599
                      litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                      temporary_code: L240804-01
                      sex: unknown
                      outcome_status: alive
                      profile_status: unindividualized
                      version: 1
                    - id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                      litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                      temporary_code: L240804-02
                      sex: unknown
                      outcome_status: alive
                      profile_status: unindividualized
                      version: 1
                    - id: 018f47a2-7582-7330-a6e3-8340a40921c1
                      litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                      temporary_code: L240804-03
                      sex: unknown
                      outcome_status: alive
                      profile_status: unindividualized
                      version: 1
                    - id: 018f47a2-7676-70b5-82a9-a375defc0b9d
                      litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                      temporary_code: L240804-04
                      sex: unknown
                      outcome_status: alive
                      profile_status: unindividualized
                      version: 1
                  initial_count_event:
                    id: 018f47a2-7768-70a2-a016-c7c12ae48550
                    litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                    event_type: initial_alive
                    delta: 4
                    occurred_at: '2026-08-04T02:30:00Z'
                    reason: 初次确认活仔
                  celebration_job:
                    id: 018f47a2-7862-7129-b21f-e415c86bbf5c
                    job_type: birth_card
                    status: queued
                    progress_percent: 0
                    retryable: true
                    version: 1
                    created_at: '2026-08-04T02:31:00Z'
                    updated_at: '2026-08-04T02:31:00Z'
                meta:
                  request_id: req_01J0BREED006
                  generated_at: '2026-08-04T02:31:00Z'
                  timezone: Asia/Shanghai
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /litters/{litter_id}/count-events:
    parameters:
      - $ref: '#/components/parameters/LitterId'
    post:
      tags: [窝次]
      operationId: createLitterCountEvent
      summary: 追加窝仔数量事件
      description: 追加数量流水；根据后补发现或关闭原因同步临时幼崽身份。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustLitterCountRequest'
            example:
              event_type: discovered
              delta: 1
              occurred_at: '2026-08-05T01:00:00Z'
              reason: 清点时后补发现一只
              new_temporary_codes:
                - L240804-05
      responses:
        '200':
          description: 数量流水和临时身份已同步
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdjustLitterCountResponse'
              example:
                data:
                  litter:
                    id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                    owner_id: 018f47a2-31dc-7e1d-9f69-3b77e4c55d11
                    origin: breeding
                    code: SY-L-20260804-01
                    breeding_plan_id: 018f47a2-4e63-7cc4-80b4-94a589917a36
                    sire_id: 018f47a2-4fe7-7823-8e1b-672ee718b014
                    dam_id: 018f47a2-50bf-7489-ab99-8fcce45842a0
                    born_at: '2026-08-04T02:30:00Z'
                    initial_alive_count: 4
                    initial_other_count: 1
                    current_managed_count: 5
                    state: nursing
                    enclosure_id: 018f47a2-6450-7f54-8851-050a3d63846f
                    dam_condition:
                      status: stable
                      notes: 精神与进食正常
                    version: 2
                    created_at: '2026-08-04T02:31:00Z'
                    updated_at: '2026-08-05T01:00:01Z'
                  count_event:
                    id: 018f47a2-7e24-78c1-b676-b5eadc5bd81a
                    litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                    event_type: discovered
                    delta: 1
                    occurred_at: '2026-08-05T01:00:00Z'
                    reason: 清点时后补发现一只
                  created_pup_identities:
                    - id: 018f47a2-7f19-7bad-b7b0-c1c4a789fab4
                      litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                      temporary_code: L240804-05
                      sex: unknown
                      outcome_status: alive
                      profile_status: unindividualized
                      version: 1
                  closed_pup_identity_ids: []
                  reconciliation:
                    initial_alive_count: 4
                    discovered_count: 1
                    deceased_count: 0
                    transferred_count: 0
                    expected_managed_count: 5
                    unindividualized_alive_count: 5
                    individualized_alive_count: 0
                    difference: 0
                    closed: false
                meta:
                  request_id: req_01J0LITTER001
                  generated_at: '2026-08-05T01:00:01Z'
                  timezone: Asia/Shanghai
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /litters/{litter_id}/wean:
    parameters:
      - $ref: '#/components/parameters/LitterId'
    post:
      tags: [窝次]
      operationId: weanLitter
      summary: 完成断奶
      description: |
        仅允许 weaning_due；对服务端计算的当前在管幼崽逐项提交生存/离舍结果并校验去向。
        断奶动作不接受或修改 profile_status，个体化进度仅由 individualize 动作推进。
        缺少当前在管身份、包含已关闭身份或笼位冲突时整体不推进状态。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeanLitterRequest'
            example:
              weaned_at: '2026-08-25T02:00:00Z'
              timezone: Asia/Shanghai
              items:
                - pup_identity_id: 018f47a2-73b3-762e-8498-07b13dc3b599
                  outcome_status: alive
                  destination_enclosure_id: 018f47a2-81bf-72b2-8990-79310bde2638
                - pup_identity_id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                  outcome_status: alive
                  destination_enclosure_id: 018f47a2-81bf-72b2-8990-79310bde2638
      responses:
        '200':
          description: 断奶已完成
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeanLitterResponse'
              example:
                data:
                  litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                  litter_state: sexing_due
                  weaned_at: '2026-08-25T02:00:00Z'
                  processed_count: 5
                  item_results:
                    - pup_identity_id: 018f47a2-73b3-762e-8498-07b13dc3b599
                      status: succeeded
                    - pup_identity_id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                      status: succeeded
                  created_task_ids:
                    - 018f47a2-83e8-795f-a0d2-04fc6630cb2f
                  version: 3
                meta:
                  request_id: req_01J0LITTER002
                  generated_at: '2026-08-25T02:00:01Z'
                  timezone: Asia/Shanghai
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /litters/{litter_id}/sex-and-separate:
    parameters:
      - $ref: '#/components/parameters/LitterId'
    post:
      tags: [窝次]
      operationId: sexAndSeparateLitter
      summary: 分性并分笼
      description: |
        仅允许 sexing_due；逐项提交性别、置信度和目标笼盒。服务端验证完整在管集合、
        异性混笼、容量和待复核安排，客户端不得提交目标状态。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SexAndSeparateRequest'
            example:
              separated_at: '2026-08-27T02:00:00Z'
              timezone: Asia/Shanghai
              items:
                - pup_identity_id: 018f47a2-73b3-762e-8498-07b13dc3b599
                  sex: male
                  sex_confidence: 0.98
                  destination_enclosure_id: 018f47a2-8620-73ef-823e-03f187be47a2
                  requires_recheck: false
                - pup_identity_id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                  sex: unknown
                  sex_confidence: 0.45
                  destination_enclosure_id: 018f47a2-8712-7863-ab76-9a6614ef3e2b
                  requires_recheck: true
      responses:
        '200':
          description: 分性分笼已完成或进入待复核安排
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Key:
              $ref: '#/components/headers/IdempotencyKey'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SexAndSeparateResponse'
              example:
                data:
                  litter_id: 018f47a2-72b3-76d8-b31f-5848c42bb577
                  litter_state: individualizing
                  separated_at: '2026-08-27T02:00:00Z'
                  processed_count: 5
                  uncertain_count: 1
                  item_results:
                    - pup_identity_id: 018f47a2-73b3-762e-8498-07b13dc3b599
                      status: succeeded
                      enclosure_stay_id: 018f47a2-8915-76c5-a87f-61ce11393494
                    - pup_identity_id: 018f47a2-748c-7d2d-a19f-ab53e616d7d8
                      status: succeeded_with_warning
                      enclosure_stay_id: 018f47a2-8a0b-7045-8d57-dda69b7e006b
                  created_task_ids:
                    - 018f47a2-8b14-7f77-98ef-b5000ec78a8b
                  version: 4
                meta:
                  request_id: req_01J0LITTER003
                  generated_at: '2026-08-27T02:00:01Z'
                  timezone: Asia/Shanghai
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /litters/{litter_id}/individualization-eligibility:
    parameters:
      - $ref: '#/components/parameters/LitterId'
    get:
      tags: [窝次, 仓鼠, 家谱]
      operationId: getLitterIndividualizationEligibility
      summary: 获取服务端个体化 eligible set
      description: |
        服务端根据窝次状态、数量账、幼崽存活状态、断奶、性别复核和有效笼位，
        计算本次必须完整转换的 pup_identity 集合。返回的 eligible_set_token 绑定
        当前 litter version 与有序身份集合；任何相关事实变化都会使旧 token 失效。
      responses:
        '200':
          description: 个体化资格集合与阻塞原因
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualizationEligibilityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 手机验证码登录后取得的访问令牌
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        写请求唯一键。唯一域为 owner_id + action_code + resource_id + key；相同规范化
        载荷返回首次结果，不同载荷返回 409 IDEMPOTENCY_PAYLOAD_MISMATCH。结果至少保留
        24 小时；confirm-birth、individualize 与分享撤销保留至对应业务记录归档。
      schema:
        type: string
        minLength: 8
        maxLength: 128
      example: 018f47a2-281b-79e2-b861-bf785ab6fba7
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 当前资源版本对应的 ETag，例如双引号包裹的整数版本。
      schema:
        type: string
        pattern: '^W?/"[0-9]+"$'
      example: '"7"'
...
    PlanId:
      name: plan_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LitterId:
      name: litter_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    IdempotencyKey:
      description: 服务端确认使用的幂等键
      schema:
        type: string
      example: 018f47a2-281b-79e2-b861-bf785ab6fba7
    IdempotencyReplayed:
      description: true 表示本次返回首次请求的已保存结果
      schema:
        type: boolean
      example: false
    ETag:
      description: 当前资源 version 的弱或强 ETag
      schema:
        type: string
        pattern: '^W?/"[0-9]+"$'
      example: '"8"'
    Location:
      description: 新建资源或异步作业的相对地址
      schema:
        type: string
      example: /v1/jobs/018f47a2-7862-7129-b21f-e415c86bbf5c
  responses:
...
    NotFound:
      description: 资源不存在、已撤销或不属于当前 owner
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 未找到指定资源
              field_errors: []
              recovery_actions: []
            meta:
              request_id: req_01J0ERR404
              generated_at: '2026-07-16T03:00:00Z'
              timezone: Asia/Shanghai
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
...
    ConfirmBirthRequest:
      type: object
      required:
        - born_at
        - initial_alive_count
        - initial_other_count
        - dam_condition
        - outcome_reason
        - timezone
      properties:
        born_at:
          type: string
          format: date-time
        enclosure_id:
          type: [string, 'null']
          format: uuid
          description: initial_alive_count > 0 时必填；N=0 时必须省略或为 null
        initial_alive_count:
          type: integer
          minimum: 0
          maximum: 100
        initial_other_count:
          type: integer
          minimum: 0
          maximum: 100
        dam_condition:
          $ref: '#/components/schemas/DamCondition'
        outcome_reason:
          type: string
          minLength: 1
          maxLength: 2000
          description: 生产结果原因；N=0 时必须明确无活仔原因
        temporary_code_prefix:
          type: [string, 'null']
          maxLength: 32
        temporary_codes:
          type: array
          uniqueItems: true
          items:
            type: string
            maxLength: 64
          description: 若提供，数量必须等于 initial_alive_count
        timezone:
          type: string
        notes:
          type: [string, 'null']
          maxLength: 4000
      additionalProperties: false
...
    ConfirmBirthResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          oneOf:
            - $ref: '#/components/schemas/ConfirmBirthLiveLitterData'
            - $ref: '#/components/schemas/ConfirmBirthNoLitterData'
          discriminator:
            propertyName: result_type
            mapping:
              live_litter: '#/components/schemas/ConfirmBirthLiveLitterData'
              no_litter_outcome: '#/components/schemas/ConfirmBirthNoLitterData'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
...
    AdjustLitterCountRequest:
      type: object
      required: [event_type, delta, occurred_at, reason]
      properties:
        event_type:
          type: string
          enum: [discovered, death, transferred_out, correction]
        delta:
          type: integer
          not:
            const: 0
          description: discovered 必须为正；death/transferred_out 必须为负；correction 非零
        occurred_at:
          type: string
          format: date-time
        reason:
          type: string
          minLength: 1
          maxLength: 2000
        new_temporary_codes:
          type: array
          uniqueItems: true
          items:
            type: string
            maxLength: 64
        affected_pup_identity_ids:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
      additionalProperties: false
...